initial commit
The data is recorded from the Zabrze-Bielszowice coal mine which is located in the area  of Poland. There are a total of 2,584 records, so the data is prominently skewed towards the non hazardous class as these are more in number. Essentially readings of energy and counts of bumps at the time of one work shift are used to classify a hazardous bump during the other shift. Hazardous bump is a seismic process with greater than 10,000 Joules of energy, and a shifting period of about 8 hours. Just consider an example, practically energy of 10,000 Joules will be the sufficient energy which is required to lift 10,000 tomatoes approximately at a height of 1m above the ground. A class represented with 0 as a result signifies a non hazardous bump, while class representing with 1 is a hazardous bump which should always be less