initial commit
Models are prepared and trained on three different algorithms; Random Forest Classifier, Logistic Regression, and Catboost classifiers after which their results and performance are analyzed. This project aims to classify the instances, on the basis of some seismic measures’ attributes, into binary categories; whether the situation leads to hazardous state or non-hazardous state. The study will propose the best suited model and algorithms for the seismic-bumps dataset.